refactor(blorktools): animation controller
